pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "azure-diagram-generator"
version = "0.1.0"
description = "Automated Azure architecture diagram generator with Lucidchart, Draw.io, and Mermaid output"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Azure Diagrammer Team"},
]
dependencies = [
    "azure-identity>=1.15.0",
    "azure-mgmt-resourcegraph>=8.0.0",
    "azure-mgmt-resource>=23.0.0",
    "azure-mgmt-network>=25.0.0",
    "azure-mgmt-compute>=30.0.0",
    "azure-mgmt-web>=7.0.0",
    "azure-mgmt-sql>=3.0.0",
    "azure-mgmt-storage>=21.0.0",
    "azure-mgmt-monitor>=6.0.0",
    "requests>=2.31.0",
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "responses>=0.24.0",
]

[project.scripts]
azure-diagrammer = "azure_diagrammer.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
